標準入力実装

+++ runner/run.py
@@ -1,25 +1,30 @@
-
-
-import sys, time, os
-if len(sys.argv) < 4:
-    print("Usage: run.py <codefile> <stdoutfile> <stderrfile> <timefile>", file=sys.stderr)
+import sys, time, os, json
+if len(sys.argv) < 2:
+    print("Usage: run.py <tmpdir>", file=sys.stderr)
     sys.exit(1)
-codefile, stdoutfile, stderrfile, timefile = sys.argv[1:5]
-with open(codefile, 'r') as f:
-    code = f.read()
+tmpdir = sys.argv[1]
+codefile = os.path.join(tmpdir, 'code.py')
+if not os.path.exists(codefile):
+    print("No code.py found in tmpdir", file=sys.stderr)
+    sys.exit(2)
+stdinfile = os.path.join(tmpdir, 'stdin.txt')
+stdoutfile = os.path.join(tmpdir, 'stdout.txt')
+stderrfile = os.path.join(tmpdir, 'stderr.txt')
+timefile = os.path.join(tmpdir, 'time.txt')
+exit_code_file = os.path.join(tmpdir, 'exit_code.txt')
+
 start = time.time()
+exit_code = 1
 try:
-    with open(stdoutfile, 'w') as out, open(stderrfile, 'w') as err:
-        import sys
-        sys.stdout = out
-        sys.stderr = err
-        try:
-            exec(code, {'__name__': '__main__'})
-        except Exception as e:
-            print(e, file=err)
-        out.flush()
-        err.flush()
+    cmd = f'{sys.executable} {codefile} < {stdinfile} > {stdoutfile} 2> {stderrfile}'
+    exit_code = os.system(cmd)
 finally:
     end = time.time()
     with open(timefile, 'w') as tf:
         tf.write(str(end-start))
+    try:
+        with open(exit_code_file, 'w') as ef:
+            ef.write(str(exit_code))
+    except Exception:
+        pass
+
